backend/requirements.txt
flask-cors
Flask
psycopg2-binary
apscheduler
gtts
spacy
python-dateutil
